policies, logs

+++ app/Http/Controllers/CorrectiveMeasuresController.php
@@ -8,6 +8,7 @@ use App\Models\Standard;
 use Illuminate\Http\Request;
 use App\Models\CorrectiveMeasure;
 use Illuminate\Support\Facades\Auth;
+use App\Facades\CustomLog;
 
 class CorrectiveMeasuresController extends Controller
 {
@@ -34,9 +35,10 @@ class CorrectiveMeasuresController extends Controller
      */
     public function create()
     {
-        $this->authorize('create',CorrectiveMeasure::class);
+        $this->authorize('create', CorrectiveMeasure::class);
+
         $standards = Standard::all();
-        $sectors = Sector::all();
+        $sectors = Sector::where('team_id', Auth::user()->current_team_id)->get();
         return view('system_processes.corrective_measures.create', compact('standards', 'sectors'));
     }
 
@@ -48,6 +50,8 @@ class CorrectiveMeasuresController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', CorrectiveMeasure::class);
+
         $correctiveMeasure = new CorrectiveMeasure();
        
         $messages = array(
@@ -92,6 +96,7 @@ class CorrectiveMeasuresController extends Controller
         $correctiveMeasure->team_id = Auth::user()->current_team_id;
 
         $correctiveMeasure->save();
+        CustomLog::info('Neusaglašenost "'.$correctiveMeasure->name.'" kreirana. Korisnik: '.\Auth::user()->name.', '.\Auth::user()->email.', '.date('d.m.Y').' u '.date('H:i:s'), 'Firma-'.\Auth::user()->current_team_id);
         $request->session()->flash('status', 'Korektivna mera je uspešno sačuvana!');
         return redirect('/corrective-measures');
     }
@@ -117,9 +122,9 @@ class CorrectiveMeasuresController extends Controller
     public function edit($id)
     {
         $corrective_measure = CorrectiveMeasure::findOrFail($id);
-        $this->authorize('update',$corrective_measure);
+        $this->authorize('update', $corrective_measure);
         $standards = Standard::all();
-        $sectors = Sector::all();
+        $sectors = Sector::where('team_id', Auth::user()->current_team_id)->get();
         return view('system_processes.corrective_measures.edit', compact('corrective_measure', 'standards', 'sectors'));
     }
 
@@ -133,7 +138,8 @@ class CorrectiveMeasuresController extends Controller
     public function update(Request $request, $id)
     {
         $correctiveMeasure = CorrectiveMeasure::findOrFail($id);
-        $this->authorize('update',$correctiveMeasure);
+
+        $this->authorize('update', $correctiveMeasure);
 
         $messages = array(
             'standard.required' => 'Izaberite standard',
@@ -173,6 +179,7 @@ class CorrectiveMeasuresController extends Controller
         }
 
         $correctiveMeasure->save();
+        CustomLog::info('Neusaglašenost "'.$correctiveMeasure->name.'" izmenjena. Korisnik: '.\Auth::user()->name.', '.\Auth::user()->email.', '.date('d.m.Y').' u '.date('H:i:s'), 'Firma-'.\Auth::user()->current_team_id);
         $request->session()->flash('status', 'Korektivna mera je uspešno izmenjena!');
         return redirect('/corrective-measures');
     }
@@ -185,7 +192,10 @@ class CorrectiveMeasuresController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('delete', CorrectiveMeasure::find($id));
+        $correctiveMeasure = CorrectiveMeasure::findOrFail($id);
         if(CorrectiveMeasure::destroy($id)){
+            CustomLog::info('Neusaglašenost "'.$correctiveMeasure->name.'" uklonjena. Korisnik: '.\Auth::user()->name.', '.\Auth::user()->email.', '.date('d.m.Y').' u '.date('H:i:s'), 'Firma-'.\Auth::user()->current_team_id);
             return back()->with('status', 'Neusaglašenost / korektivna mera je uspešno obrisana');
         }else{
             return back()->with('status', 'Došlo je do greške! Pokušajte ponovo.');
