+++ app/Http/Controllers/ManagementSystemReviewsController.php
@@ -2,15 +2,12 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Goal;
-use App\Models\PlanIp;
-use App\Models\Supplier;
-use Illuminate\Http\Request;
-use App\Models\RiskManagement;
-use App\Models\CorrectiveMeasure;
-use Illuminate\Support\Facades\Auth;
 use App\Models\ManagementSystemReview;
+use Illuminate\Http\Request;
 use App\Http\Requests\StoreManagementSystemReview;
+use App\Http\Requests\UpdateManagementSystemReview;
+use Illuminate\Support\Facades\Auth;
+use App\Facades\CustomLog;
 
 class ManagementSystemReviewsController extends Controller
 {
@@ -45,6 +42,7 @@ class ManagementSystemReviewsController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', ManagementSystemReview::class);
         return view('system_processes.management_system_reviews.create');
     }
 
@@ -56,36 +54,9 @@ class ManagementSystemReviewsController extends Controller
      */
     public function store(StoreManagementSystemReview $request)
     {
-        $msr = new ManagementSystemReview();
-
-        $year = (int)$request->year;
-        $standardId = $this::getStandard();
-
-        $msr->standard_id = $standardId;
-        $msr->year = $request->year;
-        $msr->participants = $request->participants;
-        $msr->measures_status = $request->measures_status;
-        $msr->internal_external_changes = $request->internal_external_changes;
-        $msr->customer_satisfaction = $request->customer_satisfaction;
-        $msr->monitoring_measurement_results = $request->monitoring_measurement_results;
-        $msr->resource_adequacy = $request->resource_adequacy;
-        $msr->resource_adequacy = $request->resource_adequacy;
-        $msr->checks_results_desc = $request->checks_results_desc;
-        $msr->improvement_opportunities = $request->improvement_opportunities;
-        $msr->needs_for_change = $request->needs_for_change;
-        $msr->needs_for_resources = $request->needs_for_resources;
-
-        $msr->objectives_scope = Goal::getStats($standardId, $year);
-        $msr->inconsistancies_corrective_measures = CorrectiveMeasure::getStats($standardId, $year);
-        $msr->checks_results = PlanIp::getStats($standardId, $year);
-        $msr->external_suppliers_performance = Supplier::getStats($standardId, $year);
-        $msr->measures_effectiveness = RiskManagement::getStats($standardId, $year);    
-
-        $msr->user_id = Auth::user()->current_user_id;
-        $msr->team_id = Auth::user()->current_team_id;
-        
-        $msr->save();
-
+        $this->authorize('create', ManagementSystemReview::class);
+        $msr = ManagementSystemReview::Create($request->all());
+        CustomLog::info('Zapisnik sa preispitivanja "'.$msr->year.'" kreiran. Korisnik: '.\Auth::user()->name.', '.\Auth::user()->email.', '.date('d.m.Y').' u '.date('H:i:s'), 'Firma-'.\Auth::user()->current_team_id);
         $request->session()->flash('status', 'Zapisnik je uspešno sačuvan!');
         return redirect('/management-system-reviews');
     }
@@ -111,6 +82,7 @@ class ManagementSystemReviewsController extends Controller
     public function edit(int $id)
     {
         $msr = ManagementSystemReview::findOrFail($id);
+        $this->authorize('update', $msr);
         return view('system_processes.management_system_reviews.edit', compact('msr'));
     }
 
@@ -121,51 +93,13 @@ class ManagementSystemReviewsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, int $id)
+    public function update(UpdateManagementSystemReview $request, int $id)
     {
         $msr = ManagementSystemReview::findOrFail($id);
+        $this->authorize('update', $msr);
 
-        $messages = array(
-            'participants.required' => 'Unesite učesnike',
-            'measures_status.required' => 'Unesite status mera',
-            'internal_external_changes.required' => 'Unesite promene',
-            'customer_satisfaction.required' => 'Unesite zadovoljstvo klijenata',
-            'monitoring_measurement_results.required' => 'Unesite rezultate praćenja merenja',
-            'resource_adequacy.required' => 'Unesite adekvatnost resursa'
-        );
-
-        $validatedData = $request->validate([
-            'participants' => 'required',
-            'measures_status' => 'required',
-            'internal_external_changes' => 'required',
-            'monitoring_measurement_results' => 'required',
-            'customer_satisfaction' => 'required',
-            'resource_adequacy' => 'required'
-        ], $messages);
-
-        $msr->year = $request->year;
-        $msr->participants = $request->participants;
-        $msr->measures_status = $request->measures_status;
-        $msr->internal_external_changes = $request->internal_external_changes;
-        $msr->customer_satisfaction = $request->customer_satisfaction;
-        $msr->monitoring_measurement_results = $request->monitoring_measurement_results;
-        $msr->resource_adequacy = $request->resource_adequacy;
-        $msr->resource_adequacy = $request->resource_adequacy;
-        $msr->checks_results_desc = $request->checks_results_desc;
-        $msr->improvement_opportunities = $request->improvement_opportunities;
-        $msr->needs_for_change = $request->needs_for_change;
-        $msr->needs_for_resources = $request->needs_for_resources;
-
-        $year = (int)$request->year;
-        $standardId = $this::getStandard();
-
-        $msr->objectives_scope = Goal::getStats($standardId, $year);
-        $msr->inconsistancies_corrective_measures = CorrectiveMeasure::getStats($standardId, $year);
-        $msr->checks_results = PlanIp::getStats($standardId, $year);
-        $msr->external_suppliers_performance = Supplier::getStats($standardId, $year);
-        $msr->measures_effectiveness = RiskManagement::getStats($standardId, $year);    
-        
-        $msr->save();
+        $msr->update($request->all());
+        CustomLog::info('Zapisnik sa preispitivanja "'.$msr->year.'" izmenjen. Korisnik: '.\Auth::user()->name.', '.\Auth::user()->email.', '.date('d.m.Y').' u '.date('H:i:s'), 'Firma-'.\Auth::user()->current_team_id);
         $request->session()->flash('status', 'Zapisnik je uspešno izmenjen!');
         return redirect('/management-system-reviews');
     }
@@ -178,16 +112,23 @@ class ManagementSystemReviewsController extends Controller
      */
     public function destroy(int $id)
     {
+        $this->authorize('delete', ManagementSystemReview::find($id));
+
+        $msr = ManagementSystemReview::findOrFail($id);
         if(ManagementSystemReview::destroy($id)){
+            CustomLog::info('Zapisnik sa preispitivanja "'.$msr->year.'" uklonjen. Korisnik: '.\Auth::user()->name.', '.\Auth::user()->email.', '.date('d.m.Y').' u '.date('H:i:s'), 'Firma-'.\Auth::user()->current_team_id);
             return back()->with('status', 'Zapisnik je uspešno uklonjen');
-        }else{
+        } else{
             return back()->with('status', 'Došlo je do greške! Pokušajte ponovo.');
         }
     }
 
     public function deleteApi($id)
     {
+        $msr = ManagementSystemReview::findOrFail($id);
+        $this->authorize('delete', $msr);
         ManagementSystemReview::destroy($id);
+        CustomLog::info('Zapisnik sa preispitivanja "'.$msr->year.'" uklonjen. Korisnik: '.\Auth::user()->name.', '.\Auth::user()->email.', '.date('d.m.Y').' u '.date('H:i:s'), 'Firma-'.\Auth::user()->current_team_id);
         return true;
     }
 }
